[ES-1493] Set up build job for Cordova 2048 (Android)
This build job somewhat mimics that of Cave Runner's. Specifically, add a Jenkins pipeline that will live under SDK -> Production Apps. The pattern here is that builds will be kicked off manually because the Cordova 2048 game will get its version bumped, and manually pointing to specific versions of the Skillz SDK.

There are two build job parameters of interest. The first allows a particular branch to be built. As Cordova is updated, it will have a separate branch corresponding to each major release of the Skillz SDK. The second is the build number of the job that builds and bundles the SDK. This will be for grabbing `Skillz.framework` when building the iOS version of the game.

Finally, force added the Android Studio project's gradle wrapper.