[ES_2577] Fix versioning of iOS build
I forgot that in the iOS portion of our production app builds, that scripts will attempt to retrieve a GitHub release tag for the versioning. For Android, the versioning is set in `build.gradle`. In the meantime, I've put in a kludge to default to 1.0.0 if no tag exists, and have written a story for v2 of the Cordova wrapper to have consistent versioning for both Android and iOS.